fenwick

Struct Fenwick

pub struct Fenwick<O: Op> { /* private fields */ }

A Fenwick tree (also known as a Binary Indexed Tree).

A Fenwick tree is a data structure that efficiently supports:

Point updates: Update a single element in O(log n) time

Prefix queries: Compute the fold of elements in [0, i) in O(log n) time

Range queries: (with OpSub) Compute the fold of elements in [i, j) in O(log n) time

The tree uses a generic operation Op to define the aggregation function.

Memory

A Fenwick<O> uses O(n) space where n is the length of the tree.

Examples

Creating and updating a tree with addition:

struct AddOp;
impl Op for AddOp {
    type Value = i64;
    fn identity() -> i64 { 0 }
    fn add(a: &i64, b: &i64) -> i64 { a + b }
}

let mut tree = Fenwick::<AddOp>::new(5);
tree.add(2, &10);  // Add 10 at index 2
tree.add(4, &5);   // Add 5 at index 4
let sum = tree.fold_to(..5);  // Sum of [0, 5)
assert_eq!(sum, 15);

Implementations

impl<T, O: Op<Value = T>> Fenwick<O>

pub fn new(len: usize) -> Self
where T: Clone,

Creates a new Fenwick tree with the given length.

All elements are initialized to the identity element of the operation.

Arguments

len - The number of logical elements in the tree

Examples

let tree = Fenwick::<AddOp>::new(10);

pub fn add(&mut self, index: usize, value: &T)

Adds a value to the element at the given index.

Arguments

index - The index to update (must be < length)

value - The value to add

Examples

let mut tree = Fenwick::<AddOp>::new(5);
tree.add(2, &10);

pub fn fold_to(&self, range: RangeTo<usize>) -> T

Computes the fold of all elements in the range [0, end).

Arguments

range - A range ..end specifying the prefix

Examples

let mut tree = Fenwick::<AddOp>::new(5);
tree.add(0, &1);
tree.add(1, &2);
tree.add(2, &3);
assert_eq!(tree.fold_to(..3), 6);

impl<T, O: OpSub<Value = T>> Fenwick<O>

pub fn sub(&mut self, index: usize, value: &T)

Subtracts a value from the element at the given index.

This requires the operation to support subtraction (i.e., implement OpSub).

Arguments

index - The index to update (must be < length)

value - The value to subtract

Examples

let mut tree = Fenwick::<AddOp>::new(5);
tree.add(2, &10);
tree.sub(2, &3);

pub fn fold(&self, range: Range<usize>) -> T

Computes the fold of all elements in the range [start, end).

This requires the operation to support subtraction (i.e., implement OpSub).

Arguments

range - A range [start, end) specifying the interval

Examples

let mut tree = Fenwick::<AddOp>::new(5);
tree.add(0, &1);
tree.add(1, &2);
tree.add(2, &3);
tree.add(3, &4);
assert_eq!(tree.fold(1..3), 5);  // 2 + 3